Rename all the Data and TxAuxData imports. Add deprecated module
