+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Tx.hs
@@ -79,7 +79,6 @@ import Cardano.Crypto.Hash.Class (HashAlgorithm)
 import Cardano.Ledger.Address (Addr (..), RewardAcnt (..))
 import Cardano.Ledger.Allegra.Core ()
 import Cardano.Ledger.Allegra.Tx (validateTimelock)
-import Cardano.Ledger.Alonzo.Data (Data, hashData)
 import Cardano.Ledger.Alonzo.Era (AlonzoEra)
 import Cardano.Ledger.Alonzo.PParams (
   AlonzoPParamsHKD (..),
@@ -94,6 +93,7 @@ import Cardano.Ledger.Alonzo.Scripts (
   Tag (..),
   txscriptfee,
  )
+import Cardano.Ledger.Alonzo.Scripts.Data (Data, hashData)
 import Cardano.Ledger.Alonzo.TxBody (
   AlonzoEraTxBody (..),
   AlonzoTxBody (..),
