+++ eras/conway/impl/src/Cardano/Ledger/Conway/Scripts.hs
@@ -13,13 +13,13 @@ module Cardano.Ledger.Conway.Scripts (
 where
 
 import Cardano.Ledger.Allegra.Scripts (Timelock)
-import Cardano.Ledger.Alonzo.Data (
+import Cardano.Ledger.Alonzo.Language (Language)
+import Cardano.Ledger.Alonzo.Scripts (AlonzoScript (..), isPlutusScript)
+import Cardano.Ledger.Alonzo.TxAuxData (
   AlonzoTxAuxData,
   hashAlonzoTxAuxData,
   validateAlonzoTxAuxData,
  )
-import Cardano.Ledger.Alonzo.Language (Language)
-import Cardano.Ledger.Alonzo.Scripts (AlonzoScript (..), isPlutusScript)
 import Cardano.Ledger.Babbage.Scripts (babbageScriptPrefixTag)
 import Cardano.Ledger.Conway.Era
 import Cardano.Ledger.Core
